feat: [C4R-824] fixed review + added a smaller breakpoint for galaxy fold screens
