+++ src/common/components/shape-wrapper/ShapeWrapper.tsx
@@ -5,12 +5,7 @@ interface ShapeWrapperProps {
 }
 
 const ShapeWrapper = ({ children }: ShapeWrapperProps) => {
-  return (
-    <div className="w-full h-fit bg-shape bg-repeat sm:bg-auto bg-contain">
-      <div></div>
-      {children}
-    </div>
-  );
+  return <div className="w-full h-fit bg-shape bg-repeat sm:bg-auto bg-contain">{children}</div>;
 };
 
 export default ShapeWrapper;
