+++ src/common/constants/nomenclature.constants.ts
@@ -1,0 +1,10 @@
+import { IconTypes } from '../enums/IconTypes.enum';
+
+export const LANDING_DOMAINS = [
+  { name: 'Medical', icon: IconTypes.USERS },
+  { name: 'Jurnalism', icon: IconTypes.USERS },
+  { name: 'Educație', icon: IconTypes.USERS },
+  { name: 'Inginerie', icon: IconTypes.USERS },
+  { name: 'Tech', icon: IconTypes.USERS },
+  { name: 'Advocacy', icon: IconTypes.USERS },
+];
