+++ src/common/enums/IconTypes.enum.ts
@@ -1,0 +1,3 @@
+export enum IconTypes {
+  USERS = 'users',
+}
