+++ src/pages/landing/components/domains/Domain.tsx
@@ -1,19 +1,22 @@
 import React from 'react';
 import users from '../../../../assets/icons/users.svg';
-import { IconTypes, IDomain } from './Domains';
+import { IconTypes } from '../../../../common/enums/IconTypes.enum';
+import { IDomain } from './Domains';
 
 const Domain = ({ name, icon }: IDomain) => {
   return (
-    <div className="bg-gray-100 hover:bg-yellow aspect-square flex justify-center items-center flex-col gap-y-4">
+    <div className="bg-gray-100 hover:bg-yellow xs:aspect-square aspect-auto flex justify-center items-center flex-col gap-y-4 p-3 xs:h-auto h-full">
       {icon === IconTypes.USERS ? (
         <img
-          className="xl:max-w-[8rem] md:max-w-[3rem] xs:max-w-[3.5rem] max-w-[2.8rem]"
+          className="xl:max-w-[8rem] md:max-w-[3rem] xs:max-w-[2.8rem] max-w-[1.5rem]"
           src={users}
         ></img>
       ) : (
         ''
       )}
-      <p className="font-titilliumBold sm:text-2xl lg:text-3xl text-xl">{name}</p>
+      <p className="font-titilliumBold sm:text-2xl lg:text-3xl xs:text-md text-sm text-center">
+        {name}
+      </p>
     </div>
   );
 };
