+++ src/pages/landing/components/domains/Domains.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
+import i18n from '../../../../common/configs/i18n';
+import { IconTypes } from '../../../../common/enums/IconTypes.enum';
 import Domain from './Domain';
-
-export enum IconTypes {
-  USERS = 'users',
-}
 
 export interface IDomain {
   icon: IconTypes;
@@ -11,15 +9,15 @@ export interface IDomain {
 }
 
 interface Domains {
-  title: string;
   domains: IDomain[];
 }
 
-const Domains = ({ title, domains }: Domains) => {
+const Domains = ({ domains }: Domains) => {
+  const title = i18n.t('landing:domains:title');
   return (
-    <div className="sm:px-30 sm:py-20 lg:px-40 lg:py-24 xl:px-60 xl:py-32 px-12 py-12">
+    <div className="sm:px-30 sm:py-20 lg:px-40 lg:py-24 xl:px-60 xl:py-32 xs:px-10 xs:py-10 px-5 py-5">
       <p className="font-titilliumBold sm:text-3xl lg:text-4xl text-2xl pb-8">{title}</p>
-      <div className="grid sm:gap-10 lg:gap-12 xl:gap-16 gap-8 sm:grid-cols-2 md:grid-cols-3 grid-cols-1">
+      <div className="grid sm:gap-10 lg:gap-12 xl:gap-16 gap-3 md:grid-cols-3 grid-cols-2">
         {domains.map((domain, index) => (
           <Domain name={domain.name} icon={domain.icon} key={index}></Domain>
         ))}
